Remove util.py
